fixed  add transaction dialog popup

+++ src/presentation/sections/TransactionTableSection.qml
@@ -11,7 +11,6 @@ Item {
     property int rowIndex: -1
 
     signal editTransactionClicked(var row)
-
     signal deleteTransactionClicked(int rowIndex)
 
     Dialog {
@@ -21,15 +20,17 @@ Item {
         modal: true
 
         onAccepted: {
-            root.viewModel.deleteTransaction(root.rowIndex)
-
-            if (root.rowIndex !== -1) {
+            if (root.rowIndex !== -1 && root.viewModel) {
                 root.viewModel.deleteTransaction(root.rowIndex)
                 root.rowIndex = -1
             }
         }
 
         onRejected: {
+            root.rowIndex = -1
+        }
+
+        onClosed: {
             root.rowIndex = -1
         }
     }
@@ -69,7 +70,12 @@ Item {
                 Layout.preferredHeight: 40
                 Layout.alignment: Qt.AlignVCenter
 
-                onDateChanged: function (month, year) {
+                onDateChanged: function(month, year) {
+                    if (!root.viewModel) {
+                        console.log("TransactionsPage: viewModel is undefined")
+                        return
+                    }
+
                     root.viewModel.loadData(month, year)
                     root.viewModel.setDate(month, year)
                 }
@@ -86,22 +92,46 @@ Item {
                 Layout.alignment: Qt.AlignRight
 
                 onClicked: {
-                    dialogPopup.open()
+                    if (!root.dialogPopup) {
+                        console.log("TransactionsPage: dialogPopup is undefined")
+                        return
+                    }
+
+                    root.dialogPopup.openForCreate()
                 }
             }
         }
 
         TransactionTable {
+            id: transactionTable
+
             Layout.fillWidth: true
             Layout.fillHeight: true
 
             viewModel: root.viewModel
 
-            onEditTransactionClicked: function (row) {
-                root.editTransactionClicked(row)
+            onEditTransactionClicked: function(row) {
+                if (!root.dialogPopup) {
+                    console.log("TransactionsPage: dialogPopup is undefined")
+                    return
+                }
+
+                if (row === undefined || row === null) {
+                    console.log("TransactionsPage: edit row is undefined/null")
+                    return
+                }
+
+                console.log("TransactionsPage: edit row:", JSON.stringify(row))
+
+                root.dialogPopup.openForEdit(row)
             }
 
-            onDeleteTransactionClicked: function (rowIndex) {
+            onDeleteTransactionClicked: function(rowIndex) {
+                if (rowIndex === undefined || rowIndex === null || rowIndex < 0) {
+                    console.log("TransactionsPage: invalid delete rowIndex:", rowIndex)
+                    return
+                }
+
                 root.rowIndex = rowIndex
                 deleteDialog.open()
             }
